requirements.txt
black
pytest
ruff

matplotlib
pyspark
python-dotenv
requests